+++ swagger_parser/lib/src/generator/templates/dart_dart_mappable_dto_template.dart
@@ -26,7 +26,9 @@ ${indentation(2)}const $className(${getParameters(dataClass)});
 
 ${getFields(dataClass)}
 
-${indentation(2)}static $className fromJson(Map<String, dynamic> json) => ${className}Mapper.ensureInitialized().decodeMap<$className>(json);
+${indentation(
+    2,
+  )}static $className fromJson(Map<String, dynamic> json) => ${className}Mapper.ensureInitialized().decodeMap<$className>(json);
 }
 ''';
 }
@@ -90,4 +92,7 @@ String _required(UniversalType t) =>
 
 /// return defaultValue if have
 String _defaultValue(UniversalType t) =>
-    '${t.enumType != null ? '${t.type}.${protectDefaultEnum(t.defaultValue)?.toCamel}' : protectDefaultValue(t.defaultValue, type: t.type)}';
+    '${t.enumType != null ? '${t.type}.${protectDefaultEnum(t.defaultValue)?.toCamel}' : protectDefaultValue(
+        t.defaultValue,
+        type: t.type,
+      )}';