+++ swagger_parser/lib/src/generator/templates/kotlin_typedef_template.dart
@@ -16,6 +16,8 @@ String kotlinTypeDefTemplate(
   if (type == null) {
     return '';
   }
-  return '${generatedFileComment(markFileAsGenerated: markFileAsGenerated, ignoreLints: false)}${descriptionComment(dataClass.description)}'
+  return '${generatedFileComment(markFileAsGenerated: markFileAsGenerated, ignoreLints: false)}${descriptionComment(
+    dataClass.description,
+  )}'
       'typealias $className = ${type.toSuitableType(ProgrammingLanguage.kotlin)};\n';
 }